IVOA Provenance DM - request for comments
Dear DMers,

We are requesting comments for the IVOA Provenance Data Model. The proposed
recommendation is available through its dedicated RFC web page and attached
to this email:

http://wiki.ivoa.net/twiki/bin/view/IVOA/ProvenanceRFC

Compared to last year's PR, the different classes and relations were
refined, keeping the initial base, and explaining further the connections
between classes.

There are points that are still being discussed within the group and for
which your comments would be particularly appreciated:

   - Specialized Entities: we propose in the document to have entities
   dedicated to Configuration, Description and Context for an activity. This
   also means that those entities have a primary usage, and are a priori not
   considered as Main Entities (like Data or Documents), even though they can
   be derived from those.

   - Handling of input parameters for an activity: the possibility to
   manipulate such parameters directly and independently from main entities
   has been included, but another view would be to just manipulate those
   parameters in the same way as the main entities.

We plan to organize a splinter session at the next interop in College Park
to discuss those points and the model, and hope to see you there !

Best Regards,

Mathieu Servillat for the IVOA Provenance Working Group